third_party/WebKit/LayoutTests/http/tests/security/referrer-policy-worker-has-referrer.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <meta name="referrer" content="no-referrer">
</head>
<body>
<script>
    async_test(function () {
        var workerUrl = 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/worker.php?type=report-referrer';
        var worker = new Worker(workerUrl);
        worker.onmessage = this.step_func(function (event) {
            assert_equals(workerUrl, event.data);
            this.done();
        });
    }, "Worker with no referrer policy does not inherit document's policy");
</script>
</body>
</html>
